+++ lib/logger_builder.dart
@@ -2,10 +2,11 @@
 /// Dart with good performance when disabled.
 library;
 
-export 'src/async_log_printers/async_buffered_log_printer.dart';
-export 'src/async_log_printers/async_buffered_log_printer_with_param.dart';
-export 'src/async_log_printers/async_log_printer.dart';
-export 'src/async_log_printers/async_log_printer_with_param.dart';
+export 'src/async_handlers/async_buffer.dart';
+export 'src/async_handlers/async_buffer_with_param.dart';
+export 'src/async_handlers/async_handler.dart';
+export 'src/async_handlers/async_handler_with_param.dart';
+export 'src/async_handlers/multi_publisher.dart';
 export 'src/custom_logger/entry.dart';
 export 'src/custom_logger/lazy.dart';
 export 'src/custom_logger/levels.dart';